+++ tests/tests/binding_array/mod.rs
@@ -2,3 +2,4 @@ mod buffers;
 mod sampled_textures;
 mod samplers;
 mod storage_textures;
+mod validation;